refactor(work-category): standardize DTO structure and domain types (Phase 1m)
- Move WorkCategory types from types/task/ to dedicated types/work-category/ directory
- Create CreateWorkCategoryRequest and UpdateWorkCategoryRequest DTO stubs
- Centralize work-category type exports through the domain layer index
- Update work-category service and mutation hooks to import from domain index
