+++ types/task/index.ts
@@ -1,8 +1,5 @@
 export * from './task-status';
 export * from './task-type';
-export * from './work-category';
 export * from './task';
 export * from './task-create';
 export * from './task-update';
-export * from './work-category-create';
-export * from './work-category-update';
