+++ types/work-category/index.ts
@@ -1,0 +1,3 @@
+export * from './work-category';
+export * from './work-category-create';
+export * from './work-category-update';
